Main

Description

Módulo para a realização da Tarefa 5 de LI1 em 2018/19.

Relatório

Introdução

Aqui iremos explicar o funcionamento e o raciocínio por de trás da Tarefa 5, observando as principais funções utilizadas e os seus outputs. O principal objetivo desta Tarefa é criar uma janela onde será possível observar o decorrer do Jogo e alterar o Estado do Jogo. O Principal desafio desta Tarefa foi conseguir alinhar as várias Pictures por nos definidas nos seus respetivos referenciais ( neste caso : Posicao e PosicaoGrelha).

Objetivos

Como referido anteriormente, o objetivo desta Tarefa é Obter uma imagem que represente o Estado atual do Jogo, representando no ecra todos os Disparos, Pecas, Jogadores, os seus pontos de vida e munições e todas as interações póssiveis entre estes, tentandos, igualmente, manter o aspeto do Jogo o mais interessante e atrativo possível.

Procedimento

Para atingir o objetivo, criamos um type novo que contem a informação referente ao Estado do Jogo: EstadoGloss :: ([Posicao],[Posicao],[Posicao],[Jogador],[Picture],[Disparo],Float,Mapa) Sendo o seu conteudo uma Lista que contém as Posicoes de todas as Pecas Vazias, uma Lista com as Posicoes de todas as Pecas Destrutivel, uma Lista com as Posicoes de todas as Pecas Indestrutivel, a Lista dos Jogadores, as Pictures utilizadas para representar o Estado, a lista de Disparo, o tempo do Jogo, e o Mapa que está a ser representado atualmente.

1) Desenhar o Mapa

Para desenhar o Mapa, foram usadas funções que, a partir do EstadoGloss, desenhassem cada tipo de Peca na sua Posicao. Assim, criamos funções (tiraVazM, tiraIndM e tiraDesM) que recebendo o mapa construissem as listas que iriam estar presentes no EstadoGloss

De seguida, usamos funções que colocam um tipo de Peca (vazias , dests, indests), que recebiam a lista de Pecas do EstadoGloss correspondente, e que a partir da sua Posicao iriam posicionar a Peca na localização correta, utilizando a placePos. Por exemplo a função vazias l d0 p iria percorrer toda a lista l, que contem todas as Posicoes do Mapa onde existem Pecas Vazias, e coloca-la-ia através de chamadas recursivas da função placePos. Cada imagem utilizada tem um tamanho de 50x50, sendo que o tamanho destas no Display varia conforme a dimensão do Mapa

2) Desenhar os Jogadores

Após ter o Mapa na Posicao correta, havia a necessidade de colocar os Jogadores neste. Para isso, utilizamos a translateJog, que recebe um Jogadore e o coloca no seu respetivo lugar. Uma vez que a imagem do tank é duas vezes maior do que a das Pecas, e que a Posicao dos Jogadores é do tipo PosicaoGrelha, foram feitos os devidos ajustes para que, em conjunto com a Picture do Mapa completo, os tanks estivessem alinhados e com o mesmo tamanho das restantes Pecas. Depois, foi utilizada a funcao desenhaTanks para colocar os Jogadores, chamando recursivamente a função anterior para a lista de Jogadores.

Com as Pictures do Mapa e dos Jogadores fui-nos possível cria a primeira função utilizada pelo main, desenhaEstado, que a partir das Pictures criadas pela desenhaTanks e pela desenhaMapa origina uma Picture, juntando as duas.

3) Movimentar Jogadores

Uma vez que esta Tarefa tem como objetivo criar uma imagem não estática, há necessidáde de movimentar os tanks de cada Jogador e de alterar a sua Direcao. Para isso foi utilizada a função reageEvento, que após ser pressionada uma certa tecla, altera o Estado do Jogo. Para tal, foi necessário implementar funções da Tarefa 2, para averiguar se o movimento efetuado pelo Jogador é ou não válido.

4) Criação da main

Para que o Estado fosse representado, fui ncessário criar a função main, que para além da reageEvento e desenhaEstado, utiliza também outras funções. É nesta função que se dá o load do ficheiros .bmp utilizados para desenhar o Mapa e representar os Jogadores. Defenimos o tamanho da janela, e criamos o estado inicial do Jogo: estadoInicial p = ( tiraVazM m (0,0) , tiraDesM m (0,0) , tiraIndM m (0,0) , [(Jogador (19,11) C 6 3 3),(Jogador (1,1) B 6 3 3)] , p , [] , 0 , m ) Sendo p a lista de Pictures que este recebe na main e m um exemplo de Mapa por nós defenido.

Conclusão

Concluindo, conseguimos representar os tank e o mapa de forma correta, contudo, não conseguimos utilizar a parte da Tarefa4 referente aos Disparos para completar a função reageTempoGloss. Queriamos também representar os valores referentes às vida e municoes de cada Jogador, contudo, não o conseguimos executar.

main :: IO ()

Função principal da Tarefa 5, que invoca o Jogo

type EstadoGloss = ([Posicao], [Posicao], [Posicao], [Jogador], [Picture], [Disparo], Float, Mapa)

Estado utilizado para trabalhar com o Gloss

window :: Display

Define as propriedades do Display

fr :: Int

Frame rate a que o jogo corre

estadoInicial :: [Picture] -> EstadoGloss

Estado inicial do jogo

Funções para Desenhar o EstadoGloss

desenhaEstado :: EstadoGloss -> Picture

Função que desenha o Estado do Jogo

vazias :: [Posicao] -> Dimensao -> Picture -> [Picture]

Cria uma lista das imagens a partir da lista de posições das Pecas Vazias do mapa

dests :: [Posicao] -> Dimensao -> Picture -> [Picture]

Cria uma lista das imagens a partir da lista de posições das Pecas Bloco Destrutivel do mapa

indests :: [Posicao] -> Dimensao -> Picture -> [Picture]

Cria uma lista das imagens a partir da lista de posições das Pecas Bloco Indestrutivel do mapa

desenhaMapa :: EstadoGloss -> Picture

Cria uma Picture que junta os 3 tipos de Pecas

desenhaTanks :: EstadoGloss -> Picture

Cria a Picture associada aos Tanks

translateJog :: Jogador -> Picture -> Picture

Coloca o Jogador na Posicao e roda a Picture conforme a Direcao do Jogador

Funções para Alterar o EstadoGloss

reageEvento :: Event -> EstadoGloss -> EstadoGloss

Função que altera o Estado do Jogo quando acontece um Event

moveJC :: Jogador -> [Jogador] -> [Disparo] -> Mapa -> Jogador

Verifica se pode ou nao mover-se na Direcao C

moveJB :: Jogador -> [Jogador] -> [Disparo] -> Mapa -> Jogador

Verifica se pode ou nao mover-se na Direcao B

moveJE :: Jogador -> [Jogador] -> [Disparo] -> Mapa -> Jogador

Verifica se pode ou nao mover-se na Direcao E

moveJD :: Jogador -> [Jogador] -> [Disparo] -> Mapa -> Jogador

Verifica se pode ou nao mover-se na Direcao D

moveJogC :: Jogador -> Jogador

Altera a Posicao do Jogador ou muda a Direcao para Cima, se necessário

moveJogB :: Jogador -> Jogador

Altera a Posicao do Jogador ou muda a Direcao para Baixo, se necessário

moveJogE :: Jogador -> Jogador

Altera a Posicao do Jogador ou muda a Direcao para Esquerda, se necessário

moveJogD :: Jogador -> Jogador

Altera a Posicao do Jogador ou muda a Direcao para Direita, se necessário

reageTempoGloss :: Float -> EstadoGloss -> EstadoGloss

Função que altera o Estado do jogo conforme a passagem do tempo

Funções que retiram as Pecas do Mapa

placePos :: Posicao -> Dimensao -> Picture -> Picture

Funçao que coloca uma Picture na Posicao certa

tiraVazL :: [Peca] -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Vazias de uma linha

tiraVazM :: Mapa -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Vazias de um Mapa

tiraDesL :: [Peca] -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Bloco Destrutivel de uma linha

tiraDesM :: Mapa -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Bloco Destrutivel de um Mapa

tiraIndL :: [Peca] -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Bloco Indestrutivel de uma linha

tiraIndM :: Mapa -> Posicao -> [Posicao]

Cria uma lista de Posicao das Pecas Bloco Indestrutivel de um Mapa

Funções Auxiliares

toPicture :: [Maybe Picture] -> [Picture]

Dada uma lista de Maybe Picture, transforma na lista de Picture correspondente

Funções para alterar o tamanho das Pictures

ladoMaior :: Dimensao -> Int

Calcula o maior valor da Dimensao

mapSide :: Dimensao -> Float

Calcula o tamanho de cada Peca atravez da Dimensao do Mapa

r :: Dimensao -> Float

Calcula a razao a partir das Pecas ( que têm 50x50 de tamanho)

resized :: Picture -> Dimensao -> Picture

Altera o tamanho da Picture

Exemplos

m :: Mapa

Exemplo de Mapa

d0 :: Posicao

Dimensao do Mapa exemplo