Tarefa1_2018li1g035

Description

Este módulo define funções comuns da Tarefa 1 do trabalho prático.

Testes

testesT1 :: [Instrucoes]

Testes unitários da Tarefa 1.

Cada teste é uma sequência de Instrucoes.

Funções principais da Tarefa 1.

instrucao

Arguments

:: Instrucao
A Instrucao a aplicar.

-> Editor
O Editor anterior.

-> Editor
O Editor resultante após aplicar a Instrucao.

Aplica uma Instrucao num Editor.

Move - move numa dada Direcao.

MudaTetromino - seleciona a Peca seguinte (usar a ordem léxica na estrutura de dados), sem alterar os outros parâmetros.

MudaParede - muda o tipo de Parede.

Desenha - altera o Mapa para incluir o Tetromino atual, sem alterar os outros parâmetros.

mudaTetro :: Tetromino -> Tetromino

Muda o tipo de Tetromino

mudaPar :: Parede -> Parede

Muda o tipo de Parede

rodaDir :: Direcao -> Direcao

Muda a Direcao para a proxima

tamanhoTetro :: Tetromino -> Int

Dá o tamanho do Tetromino

dirMat :: Tetromino -> Direcao -> Matriz Bool

Poe o Tetromino na Direcao correta

desMapa :: Editor -> (Int, Int) -> Editor

Recursiva da fazPecaL para as linhas

fazPecaL :: Matriz Bool -> Editor -> (Int, Int) -> Editor

Recursiva da fazPeca para uma linha

fazPeca :: Matriz Bool -> Editor -> (Int, Int) -> Editor

Muda uma Peca

posiMat :: Matriz Bool -> Posicao -> Bool

Ve onde vai comecar o Tetromino

instrucoes

Arguments

:: Instrucoes
As Instrucoes a aplicar.

-> Editor
O Editor anterior.

-> Editor
O Editor resultante após aplicar as Instrucoes.

Aplica uma sequência de Instrucoes num Editor.

NB: Deve chamar a função instrucao.

newdimensao :: Dimensao -> Dimensao

Cria um Mapa inicial com Paredes nas bordas e o resto vazio.

dimensaoToPosicoesLinha :: Dimensao -> [(Int, Int)]

Cria o numero de Colunas da Matriz

dimensaoToPosicoesMatriz :: Dimensao -> [[(Int, Int)]]

Cria o numero de Linhas da Matriz

linhaPosParaPecas :: Dimensao -> [(Int, Int)] -> [Peca]

Preenche as linhas com Peca

matrizPosParaPecas :: Dimensao -> [[(Int, Int)]] -> Mapa

Preenche a Matriz com as varias Linhas

mapaInicial :: Dimensao -> Mapa

Cria a Matriz a partir da Dimensao

editorInicial

Arguments

:: Instrucoes
Uma sequência de Instrucoes de forma a poder calcular a dimensaoInicial e a posicaoInicial.

-> Editor
O Editor inicial, usando a Peca I Indestrutivel voltada para C.

Cria um Editor inicial.

NB: Deve chamar as funções mapaInicial, dimensaoInicial, e posicaoInicial.

mudaPos :: Instrucoes -> Posicao -> Posicao

Altera a Posicao apos a Instrucao Move

constroi

Arguments

:: Instrucoes
Uma sequência de Instrucoes dadas a um Editor para construir um Mapa.

-> Mapa
O Mapa resultante.

Constrói um Mapa dada uma sequência de Instrucoes.

NB: Deve chamar as funções Instrucoes e editorInicial.

tiraMapa :: Editor -> Mapa

Retira o Mapa de um Editor