Tarefa2_2018li1g035

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

procuraJogador :: [Jogador] -> Int -> Jogador

Substitui um jogador na Lista

repoeJogador :: [Jogador] -> Jogador -> Int -> [Jogador]

Substitui um jogador na Lista

removeJogador :: [Jogador] -> Int -> [Jogador]

Remove um Jogador da lista

comecoDoTiro :: PosicaoGrelha -> Direcao -> PosicaoGrelha

Coloca o Tiro na Grelha

atualizaListaDisparos :: [Disparo] -> Jogador -> Int -> Jogada -> [Disparo]

Atualiza o Disparo conforme o tipo de Disparo

atualizaPosicao :: Jogador -> Jogada -> Jogador

Atualiza a Posicao de um Jogador

atualizaDirecao :: Jogador -> Jogada -> Jogador

Atualiza a Direcao de um Jogador

atualizaLaser :: Jogador -> Jogada -> Jogador

Atualiza o numero de Municoes de Laser

atualizaChoque :: Jogador -> Jogada -> Jogador

Atualiza o numero de Municoes de Choques

atualizaPosJogSePos :: Jogador -> Jogada -> Jogador

Atualiza a Posicao quando o jogador esta na mesma direcao

posicoesOcupadas :: Jogador -> [Posicao]

Cria uma lista com as posicoes ocupadas pelo Jogador

posicoesJogadores :: [Jogador] -> [[Posicao]]

Usa a posicoesOcupadas para vários Jogadore

comparaPosicao :: Posicao -> Posicao -> Bool

Compara duas posicoes e verifica se sao iguais

comparaPosicaoComLista :: Posicao -> [Posicao] -> Bool

Compara uma Posicao com uma lista de Posicoes e verifica se sao iguais

comparaPosicaocomMatriz :: Posicao -> [[Posicao]] -> Bool

Compara uma Posicao com as de uma Matriz e verifica se sao iguais

verificaPosicoes :: [Posicao] -> [[Posicao]] -> Bool

Verifica se alguma das Posicoes da lista estao na Matriz

veSeHaChoques1 :: Disparo -> [Int]

Cria uma Lista com o jogadores ativou o choque

veSeHaChoques2 :: [Disparo] -> [Int]

Cria uma lista com os jogadores que ativaram os choques

veOJogador :: [Jogador] -> Int -> Jogador

Tira um jogador da lista de Jogadores

veOJogador2 :: [Jogador] -> [Int] -> [Jogador]

Cria uma lista com os jogadores que foram retirados

posicaoToListadePosicoesChoque :: Jogador -> [Posicao]

Cria uma Lista das Posicoes onde o Choque e efetivo

posicaoToListadePosicoesChoque2 :: [Jogador] -> [[Posicao]]

Cria uma lista de Posicoes onde o Choque e efetivo por cada jogador

verificaMapa :: Mapa -> Jogador -> Jogada -> Bool

Verifica se tem paredes

verificaFinal :: Jogador -> [Jogador] -> [Disparo] -> Mapa -> Jogada -> Jogador

Verifica se pode ou nao mover-se na Direcao

jogada :: Int -> Jogada -> Estado -> Estado

Efetua uma jogada