Tarefa3_2018li1g035

Description

Este módulo define funções comuns da Tarefa 3 do trabalho prático.

Relatório

Introdução

Começando, iremos explicar todo o procedimento da tarefa 3 , bem como o método usado por nós para a conclusão da mesma ,e o quão útil esta é para o jogo. Iremos também interpretar os resultados obtidos após a compressão/descompressão do estado do jogo , e também o porquê desses mesmos resultados. Sendo assim , o principal desafio desta tarefa foi criar uma maneira o mais eficiente possível de comprimir e logo após descomprimir o estado do jogo recorrendo a funções que permitissem tal coisa.

Objetivos

$obs O principal objetivo desta tarefa , como foi anteriormente referido , é criar uma fórmula que permita comprimir e descomprimir o estado do jogo da maneira mais eficaz possível , de forma a reduzir o espaço utilizado na gravação do estado atual do jogo. Sendo assim , e como nos foi pedido , iniciamos por comprimir os estado do jogo para só depois descomprimir usando o mesmo método de pensamento.

Procedimento

O método de pensamento usado foi dividir a estratégia de compressão por 3 partes : a compressão do mapa , a compressão da lista dos jogadores e compressão da lista de disparos. Assim sendo , a estratégia usada foi a seguinte :

Compressão do estado 

a - Compressão do mapa

No âmbito da compressão do mapa , começamos por remover do mapa as paredes laterais , ou seja , as chamadas bordas para reduzir ao número de caratéres a usar na compressão, isto recorrendo ao uso de 2 auxiliares. De seguida, usamos o v para identificar a peça Vazia, o d para a peça 'Bloco Destrutivel' e o i para a peça 'Bloco Indestrutivel'. Por fim usamos o caratér . no final de cada linha para identificar as diferentes linhas do mapa. Exmplificando :

Input - listaMapa (novoMapa2 (mapaAux3 [[Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel]]))

Output - "vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv."

b - Compressão da lista de jogadores

No que toca à compressão da lista de jogadores , o estratagema usado foi primeiramente iniciar a lista de jogadores com o caratér j para destinguir a mesma do mapa. Depois , usou-se o caratér . para distinguir os jogadores e todos os inteiros foram separados por virgúlas para podermos converter em String usando a função Show, e por último usou-se o caratér * para distinguir a direção do jogador dos inteiros. Sendo que se converteu jogador a jogador e string e tudo foi concatenado no fim. Assim sendo, recorremos ao seguinte exemplo para mostrar o funcionamento deste método:

Input - listaJog [(Jogador (1,2) B 1 1 1),(Jogador (6,6) C 1 1 1)]

Output - "1,2,1,1,1,*B.6,6,1,1,1,*C."

Nota : o caratér j apenas foi adicionado na concatenação das 3 strings comprimidas , daí não aparecer no output.

c - Compressão da lista de disparos

Quanto à lista de disparos a estratégia usada foi semelhante à estretégia usada na lista de jogadores , começamos assim por separar esta string da string da lista de jogadores recorrendo ao caratér t sendo que também foi apenas adicionado na concatenação final das strings. Quanto a identificação do tipo de disparo , usou-se o caratér c para o disparo de canhão , o caratér C para o disparo de choque e o disparo de laser foi identificado pelo caratér L. A restante estratégia , isto é , a estratégia usada para a separação dos inteiros , ou seja , o identificador do jogador , a posição e o tempo , foi a mesma usada na lista de jogadores , assim como o identificador da direçao. Como exemplo temos :

Input - listaDisp [(DisparoCanhao 0 (2,2) B),(DisparoChoque 2 2),(DisparoLaser 1 (3,1) D)]

Output - "c0,2,2,*B.C2,2*.L1,3,1,*D."

Descompressão do estado 

No âmbito da descompressão a estratégia usada foi a mesma tanto na descompressão do mapa , como na das listas de jogadores e de disparos , foi recorrer ao take e drop while usando os caratéres de distinção falados anteriormente como pontos de "paragem" e usando o descString para passar as strings que representam inteiros para inteiros.No fim , concatenou-se tudo num só estado , sendo que as 3 listas apesar de comprimidas com igual método , foram descomprimidas de forma separada. Assim sendo , apresentaremos agora os exemplos de descompressão, tanto do mapa,como dos jogadores e como dos dosparos.

a - Descompressão do mapa 

Input - descMapaFinal "vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.j1,2,1,1,1,*B.6,6,1,1,1,*C.tc0,2,2,*B.C2,2*.L1,3,1,*D."

Output - [[Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Vazia,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia,Vazia ,Vazia,Bloco Indestrutivel] ,[Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel,Bloco Indestrutivel,Bloco Indestrutivel ,Bloco Indestrutivel]]

b - Descompressão da lista de jogadores

Input - descJogFinal "vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.j1,2,1,1,1,*B.6,6,1,1,1,*C.tc0,2,2,*B.C2,2*.L1,3,1,*D."

Output - [Jogador {posicaoJogador = (1,2), direcaoJogador = B, vidasJogador = 1, lasersJogador = 1, choquesJogador = 1},Jogador {posicaoJogador = (6,6), direcaoJogador = C, vidasJogador = 1, lasersJogador = 1, choquesJogador = 1}]

c - Descompressão da lista de disparos 

Input - descTirosFinal "vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.vvvvvvvv.j1,2,1,1,1,*B.6,6,1,1,1,*C.tc0,2,2,*B.C2,2*.L1,3,1,*D."

Output - [DisparoCanhao {jogadorDisparo = 0, posicaoDisparo = (2,2), direcaoDisparo = B},DisparoChoque {jogadorDisparo = 2, tempoDisparo = 2}, DisparoLaser {jogadorDisparo = 1, posicaoDisparo = (3,1), direcaoDisparo = D}]

Conclusão

$conc Em suma , os resultados obtidos foram positivos sendo que todos os testes feitos não apresentaram erros , tendo obtido bons resultados no site de testes disponiblizado pelos professores , mas no que toca a termos de avaliação não se obteu o total de nota da tarefa , o que prova que se poderia ter testado mais vezes e de maneiras diferentes porque eventualmente não esteve tão eficaz como desejado, mas ,ainda assim , em termos de percentagem de avaliação foi uma avaliação positiva sendo que tivemos uma boa classificação.

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3.

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> String

Comprime um Estado para formato textual.

NB: A função show representa um Estado num formato textual facilmente legível mas extenso.

NB: Uma boa solução deve representar o Estado dado no mínimo número de caracteres possível.

dirToChar :: Direcao -> Char

Funcao Auxiliar para a comprime --

mapaAux1 :: [a] -> [a]

Funcao auxiliar para comprimir o Mapa

mapaAux2 :: [a] -> [a]

Funcao auxiliar para comprimir o Mapa

mapaAux3 :: Mapa -> Mapa

Funcao auxiliar para comprimir o Mapa

novoMapa :: Mapa -> Mapa

Funcao auxiliar para comprimir o Mapa

removeUltima :: Mapa -> Mapa

Funcao auxiliar para comprimir o Mapa

novoMapa2 :: Mapa -> Mapa

Funcao auxiliar para comprimir o Mapa

pecaToString :: Peca -> String

Comprime uma Peca para uma String

pecasToString :: [Peca] -> String

Comprime uma lista de Pecas para uma String

pecasToString1 :: [Peca] -> String

Termina a String com "."

mapaToString :: Mapa -> String

Comprime um Mapa para uma String

listaMapa :: Mapa -> String

Termina a String do Mapa

jogToString :: Jogador -> String

Comprime um jogador

listaJog :: [Jogador] -> String

Comprime uma lista de Jogadores

dispToString :: Disparo -> String

Comprime um Disparo

listadToString :: [Disparo] -> String

Comprime uma lista de Disparos

listaDisp :: [Disparo] -> String

Termina a String da Lista de Disparos

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

NB: A função comprime é válida de for possível recuperar o Estado utilizando a função descomprime, i.e.:

descomprime . comprime = id

NB: Esta propriedade é particularmente válida para a solução pré-definida:

read . show = id

descString :: String -> Int

Descomprime para inteiro

descSInt :: String -> [Int]

Descomprime uma String para uma lista de Ints

stringToDir :: String -> Direcao

Transforma uma String numa Direcao

descPeca :: Char -> Peca

Transforma um Char num tipo de Pecas

descListaPeca :: String -> [Peca]

Transforma uma String numa Lista de Pecas

descMapa :: String -> Mapa

Descomprime o Mapa sem as bordas

descMapa0 :: [Peca] -> [Peca]

Adiciona as bordas laterais a uma linha

descMapa1 :: Mapa -> Mapa

Adiciona as bordas laterais ao Mapa

descMapa2 :: Int -> [Peca]

Cria uma linha de Blocos Indestrutiveis

descMapa4 :: Mapa -> Mapa

Adiciona a primeira e ultima linha de blocos Indestrutiveis ao mapa

descMapaFinal :: String -> Mapa

Descomprime o Mapa

descToPlayer :: [Int] -> String -> Jogador

Descomprime a String de um Jogador

descJog :: String -> Jogador

Reduz a String para a String de um Jogador

descJog1 :: String -> [Jogador]

Descomprime uma Lista de Jogadores

descJogFinal :: String -> [Jogador]

Reduz uma String para uma Lista de Jogadores

tiroFromString :: [Int] -> String -> Disparo

Descomprime um Disparo

tiroFromString1 :: String -> Disparo

Reduz uma String para um Disparo

descTiros :: String -> [Disparo]

Descomprime uma lista de Disparos

descTirosFinal :: String -> [Disparo]

Reduz uma Lista de Resultados

jog1 :: Jogador

Exemplo de Jogador1

jog2 :: Jogador

Exemplo de Jogador2

jog3 :: Jogador

Exemplo de Jogador3

jog4 :: Jogador

Exemplo de Jogador4

dis1 :: Disparo

Exemplo de Disparo Canhao

dis2 :: Disparo

Exemplo de Disparo Choque

dis3 :: Disparo

Exemplo de Disparo Choque

dis4 :: Disparo

Exemplo de Disparo Lase

estado :: Estado

Exemplo de Estado

m :: Mapa

Mapa exemplo