Tarefa4_2018li1g035

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

Lasers

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

verificaSeELaser :: Disparo -> Bool

Efeito dos lasers no mapa.

Verifica se um Disparo é Laser, returnando False caso contrário

removeSemSerLaser :: [Disparo] -> [Disparo]

Retira da lista de Disparo os que não são Laser

posicoes1laser :: Disparo -> Mapa -> [Posicao]

Cria a Lista de Posicoes afetadas por um Lasers

posicoesLasers :: [Disparo] -> Mapa -> [Posicao]

Cria uma Lista das Posicoes afetadas por todos os Disparos Laser

posicoesOcupadas1jog :: Jogador -> [Posicao]

Efeito dos lasers nos jogadores.

Cria a lista de Posicoes ocupadas por um Jogador

atualizaVidasLaser :: [Jogador] -> [Disparo] -> Mapa -> [Jogador]

Diminui a vida dos Jogadores que são atingidos por um Laser

posicoesexatas1laser :: Disparo -> Mapa -> [Posicao]

Efeito dos lasers nos disparos e sua atualização.

Cria a lista de Posicoes afetadas pelo Laser, parando no primeiro Bloco Indestrutivel que atingir

posicoesLasers2 :: [Disparo] -> Mapa -> [Posicao]

Cria a lista de Posicoes afetadas por todos os Lasers ativos

comparaCanhaocomLasers :: Disparo -> [Posicao] -> Bool

Compara as Posicoes de um Disparo Canhao

removeCanhaoSeLaser :: [Disparo] -> Mapa -> [Disparo]

Remove um Disparo Canhao caso seja atingido por um Laser

removeselaser :: [Disparo] -> [Disparo]

Retira um Disparo Laser da Lista de Disparos

Canhões

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

verificaPosIgual :: Posicao -> Posicao -> Bool

Funções de ajuda

Verifica se duas Posicoes são iguais

verificaPosIgual1 :: Posicao -> [Posicao] -> Bool

Verifica se uma Posicao esta presente numa Lista de Posicoes

verificaPosFinal :: [Posicao] -> [Posicao] -> Bool

Verifica se alguma das Posicoes da primeira lista esta presente na segunda

verificaSeECanhao :: Disparo -> Bool

Retirar vidas aos jogadores se houver choque contra algum canhão.

Verifica se um Disparo é Canhao

canhaoToPos :: Disparo -> Posicao

Retira a Posicao do Canhao

soCanhoes :: [Disparo] -> [Disparo]

Retira da Lista de Disparos todos os que não são Disparos Canhao

posCanhoes :: [Disparo] -> [Posicao]

Cria a Lista de Posicoes afetadas pelos Canhoes

posicoesProibidas :: Jogador -> [Posicao]

A partir de um Jogador cria uma lista de Posicoes proibidas

atualizaVidaDe1Jog :: Jogador -> [Disparo] -> Jogador

Retira uma vida a um Jogador que é atingido por um Disparo Canhao

atualizaVidasJog :: [Jogador] -> [Disparo] -> [Jogador]

Retira uma vida aos Jogadores que forem atingidos por Disparos Canhao

disparoPos :: Disparo -> [Posicao]

Atualizar o mapa se houver choque de algum canhão.

Cria a Lista de Posicoes afetadas pelo Disparo Canhao num dado local

disparosPos :: [Disparo] -> [Posicao]

Cria a Lista de Posicoes afetadas por vários Disparos Canhao

atualizaMapa :: [Posicao] -> Mapa -> Mapa

Remove as Pecas Bloco Destrutiveis quando são atingidas, e substitui-as por Pecas Vazias

comparaCanhoes :: Disparo -> Disparo -> Bool

Explosões dos tiros de canhão.

Ve se 2 Disparos Canhao estao na mesma PosicaoGrelha

comparaCanhoes2 :: Disparo -> [Disparo] -> Bool

Ve se um Disparo canhao esta na mesma posicao de outro da lista

listaColiEntCanhoes :: [Disparo] -> [Disparo] -> [Disparo]

Compara duas listas de Disparos Canhao

verificaColiTanques :: Disparo -> Jogador -> Bool

Verifica se o Disparo Canhao esta a colidir com algum Jogador

verificaColiTanques2 :: Disparo -> [Jogador] -> Bool

Verifica se o Disparo Canhao esta a colidir com algum dos Jogadores da Lista

verificaColiParedes :: Disparo -> Mapa -> Bool

Verifica se o Disparo Canhao esta a colidir com algum Bloco Indestrutivel ou Bloco Destrutivel

atualizaverificando :: [Disparo] -> [Jogador] -> Mapa -> [Disparo]

Atualiza a lista de Disparos, verificando atraves das funcoes anteriores

atualizaPosicoesCanhao :: Disparo -> Disparo

Atualiza a PosicaoGrelha do Disparo Canhao

atualizaPosicoesCanhao2 :: [Disparo] -> [Disparo]

Atualiza a PosicaoGrelha da lista de Disparos

Choques

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

verificaSe0 :: Disparo -> Bool

Funçoes que removem os choques que estejam com a contagem dos ticks a 0.

Verifica se o t de um Disparo Choque esta a 0

removeTicks0 :: [Disparo] -> [Disparo]

Remove um Disparo Choque da lista de Disparos, se este tiver t==0

atualizaSeChoque :: Disparo -> Disparo

Funções que atualizam o tempo dos choques

Decrementa o t de um Disparo, caso seja Laser

atualizaChoques :: [Disparo] -> [Disparo]

Atualiza os Disparos Choques da lista de Disparos

Exemplos

jog1 :: Jogador

Exemplo de Jogador1

jog2 :: Jogador

Exemplo de Jogador2

jog3 :: Jogador

Exemplo de Jogador3

jog4 :: Jogador

Exemplo de Jogador4

dis1 :: Disparo

Exemplo de Disparo Canhao

dis2 :: Disparo

Exemplo de Disparo Choque

dis3 :: Disparo

Exemplo de Disparo Choque

dis4 :: Disparo

Exemplo de Disparo Laser

estado :: Estado

Exemplo de Estado1

estado2 :: Estado

Exemplo de Estado2

estado3 :: Estado

Exemplo de Estado3

m1 :: Mapa

Exemplo de Mapa1

m2 :: Mapa

Exemplo de Mapa2

m3 :: Mapa

Exemplo de Mapa3