Tarefa6_2018li1g035

Description

Este módulo define funções comuns da Tarefa 6 do trabalho prático.

Relatório

Introdução

Inicialmente, será explicado o conceito da tarefa 6 e estratégia que nós adotamos para a sua realização e ,por fim ,a sua utilidade. Iremos falar dos resultados obtidos após a programação deste bot , e também o porquê de ter sido realizada da forma que foi. Sendo assim , o principal desafio desta tarefa foi criar um bot o mais eficiaz possível , isto é , programá-lo de modo a conesguir o máximo de pontos possíveis ao fim de 200 ticks , duração de um jogo.

Objetivos

O principal objetivo desta tarefa era contruir um bot de modo a que ao fim de 200 ticks ele tivesse mais pontuação que todos os outros jogadores em jogo. A fim deste objetivo utilizamos a seguinte estretégia : primeiro ter em conta o que o bot poderia fazer com os tiros disponíveis pois , afinal , a nosso ver é a melhor maneira de obter uma boa pontuação e só depois pensar no que ele poderia fazer com os movimentos de uma forma equilibrada e não disparar para qualquer Direcao. Deu-se assim prioridade aos disparos em deterimento dos movimentos.

Procedimento

O método de pensamento usado foi dividir a estratégia de compressão por 2 partes : a possível utilização de disparos como prioridade e ,logo após, a possível utilização dos movimentos. Dentro dos disparos , e sendo o disparo mais forte , inicialmente defenimos que se houver justificção para tal deveria-se usar o laser , visto que é o mais propício à obtenção de pontos , depois e canhão e por último o choque. Quanto a possiveis movimentos e se não se justificar um disparo , defenimos como prioritário o avançar se tiverem reunidas condições para tal acontecimento e só depois mudar a sua direção.

Utilização de disparos 

a - Utilização do laser

Sendo este o disparo mais temível , foi à utilização do mesmo a que demos prioridade , sendo que , segundo o que nós achamos , ele irá jogar um "Dispara Laser" se houver algum jogador no seu campo de disparo , isto é , se até o laser se parar de propagar (até encontrar uma parede indestrutivel) algum jogador estiver nessa linha de disparo, o laser irá ser utilizado , e a outra ocasião é se ele encontrar paredes indestrutiveis nesse mesmo campo , sendo que apenas o fará nestas 2 ocasiões se e só se tiver o contador de laser superior a 0. Assim o defenimos porque o tirar vidas a jogadores e o destruir paredes destrutiveis contam 4 e 1 pontos respetivamente para a pontuação total do jogador, o que significa obtenção de pontos e sendo o laser o disparo mais forte nenhum outro disparo o deterá.

b - Utilização do canhão

No que toca ao disparo de canhão que tem como grande vantagem ser infinito , a nosso ver é a 2ª melhor arma do jogo pois para além do que foi dito , também possiblita o jogador de tirar vidas ao adversário e destruir paredes destrutiveis , assim sendo este disparo irá ser usado se algum jogador se encontrar a até 4 posições de distância da linha de disparo do bot , isto porque , estando ele perto, é muito mais dificil de escapar à bala. Se alguma bala de canhão estiver a 4 ou menos posições do jogador , isto irá permitir destruir a bala de canhão e preservar as vidas do bot , relembrando que cada uma vale 4 pontos. Por último, será usada se existir alguma parede destrutivel nos blocos à frente do jogador , o que para além de permitir ganhar pontos, permite um possível avanço do jogador se necessário.

c - Utilização do Choque

A utilização do choque também tem a sua vantagem , e esta é o facto de imobilizar algum jogador no raio de 3 posições do jogador , o que permite que represente menos perigo pois há menos hipóteses de o bot ser atingido. Sendo assim , se não se justificar o uso de um disparo canhão ou um disparo laser e se o bot encontrar algum jogador no raio de 3 posições, irá usar o choque , isto se o contador de choques for superior a 0.

Utilização dos Movimentos

a - Movimentação na direção atual. 

A movimentação também é importante pois permite "fugir" de possíveis disparos e preservar vidas. Sendo assim, o bot irá movimentar-se para a frente se vier alguma bala numa direção perpendicular à sua direção atual , para este poder escapar à bala ou pelo menos tentar e também se for possível movimentar-se para a frente , isto é , se não existir alguma parede a impossibilitá-lo.

b - Mudança de direção.

O bot irá mudar a sua direção se estiver impossibilitado de se mover para a frente, isto porque se tiver alguma parede indestrútivel à frente será impossível a a sua movimentação nesse sentido e também se vier alguma bala na sua direção ou da direção oposta à que ele está para permitir uma possível fuga e salvaguardar as suas vidas.

Não jogar

A última hipótese e, a nosso ver, a com menos utilidade, é o jogador não jogar , ou seja, não fazer nada. Mas se por algum motivo não fôr possível a realização de alguma jogada anteriormente referida, assim o fará.

Conclusão

Concluido , apesar de achamos que esta foi a tarefa mais desafiante , em termos de resultados obtidos foi talvez uma das menos eficazes uma vez que não foi testada o suficiente , mas , mesmo assim , o bot responde e executa sempre uma jogada , sendo que talvez nao seja a mais indicada para o tempo , ou seja , não é um bot totalmente "perfeito".

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

procuraJogador :: [Jogador] -> Int -> Jogador

Procura um jogador na Lista

removeJogador :: [Jogador] -> Int -> [Jogador]

Remove um Jogador da lista

verificaPosIgual :: Posicao -> Posicao -> Bool

Verifica se duas Posicoes sao iguais

verificaPosIgual1 :: Posicao -> [Posicao] -> Bool

Verifica se uma Posicao é igual a alguma de uma lista

verificaPosFinal :: [Posicao] -> [Posicao] -> Bool

Verifica se alguma posicao das duas listas esta presente na outra

Utilização do Laser

verificaSeTemLaser :: Jogador -> Bool

Verifica se um Jogador não tem Lasers

posicoes1laser :: Jogador -> Mapa -> [Posicao]

Cria a lista de Posicoes afetadas pelo Laser

posicoesOcupadas1jog :: Jogador -> [Posicao]

Cria a lista de Posicoes ocupadas por um Jogador

posicoesOcupadasJog :: [Jogador] -> [Posicao]

Cria a lista de Posicoes ocupadas pelos Jogadores

verificaSeTemBDes :: [Posicao] -> Mapa -> Bool

Verifica se existem Blocos Destrutiveis numa Lista de Posicoes

verificaSeJogaLaser :: Int -> [Jogador] -> Mapa -> Bool

Verifica se um Jogador utilizou um Lazer

verificaSeJogaLaserAux :: Int -> [Jogador] -> Mapa -> Bool

Verifica se um Laser vai atingir um Jogador

Utilização do Canhão

verificaParedesFrente :: Jogador -> Mapa -> Bool

Verifica se um Jogador tem Blocos Destrutiveis à sua frente

verificaSeECanhao :: Disparo -> Bool

Verifica se um Disparo é um Disparo Canhao

soCanhoes :: [Disparo] -> [Disparo]

Retira da lista de Disparos os que nao sao Canhao

canhaoToPos :: Disparo -> Posicao

Retira a Posicao do Disparo

posCanhoes :: [Disparo] -> [Posicao]

Retira a Posicao de varios Disparos Canhao

posicoesdeperigoCanhao :: Jogador -> [Posicao]

Cria a lista de Posicoes de Perigo

posicoesdeperigoTanques :: Jogador -> [Posicao]

Cria a lista de Posicoes onde um Disparo Canhao vai afetar o Jogador

verificaSeJogaCanhao :: Int -> [Jogador] -> [Disparo] -> Mapa -> Bool

Verifica se um Jogador vai ser atingido um um Disoari

Utilização do Choque

verificaSeTemChoque :: Jogador -> Bool

Verifica se o Jogador tem algum Disparo Choque

posicaoToListadePosicoesChoque :: Jogador -> [Posicao]

Cria a lista de Posicoes afetadas pelo Disparo Choque

verificaSeJogaChoque :: Int -> [Jogador] -> Bool

Verifica se vai utilizar um Disparo Choque

Possibilidade de Avançar

verificaSePodeAvancar :: Jogador -> Mapa -> Bool

Verifica se o Jogador pode avançar na sua Direcao

verificaCanhoesDirOposta :: Jogador -> Posicao -> Bool

Verifica se algo esta na Direcao do Jogador

verificaCanhoesDirOpostaFinal :: Jogador -> [Posicao] -> Bool

Verifica se ha algo na Direcao do Jogador

avanca :: Jogador -> Jogada

Movimenta o Jogador na sua Direcao

verificaSeAvanca :: Int -> [Jogador] -> [Disparo] -> Mapa -> Bool

Verifica se deve ou nao avancar

Possibilidade de Mudar de Direção

mudaDir :: Jogador -> Jogada

Muda a Direcao do Jogador

verificaCanhoesMesmaDir :: Jogador -> Posicao -> Bool

Verifica se algo esta na Direcao do Jogador

verificaCanhoesMesmaDirFinal :: Jogador -> [Posicao] -> Bool

Verifica se ha algo na Direcao do Jogador

verificaSeMudaDir :: Int -> [Jogador] -> [Disparo] -> Mapa -> Bool

Verifica se deve ou nao mudar de Direcao